requirements.txt
flask
python-dotenv
requests
azure-storage-blob
azure-ai-translation-document
